fix: random number 값 받는 로직 수정
- index.js에서 random number 값 관리
